+++ packages/server/src/resourceProviders/postgres/search/index.ts
@@ -14,62 +14,13 @@ import {
   findSearchParameter,
   SearchParameterResource,
   SearchParameterResult,
-  searchParameterToTableName,
   parametersWithMetaAssociated,
   deriveResourceTypeFilter,
   deriveLimit,
 } from "../../utilities/search/parameters.js";
 
 import { deriveSortQuery } from "./sort.js";
-import { PARAMETER_FILTERS } from "./parameters/index.js";
-
-export function buildParameterSQL(
-  ctx: FHIRServerCTX,
-  parameter: SearchParameterResource,
-  index: number,
-  values: unknown[],
-  columns: string[] = ["DISTINCT(r_version_id)"]
-): { index: number; query: string; values: unknown[] } {
-  const searchParameter = parameter.searchParameter;
-  const search_table = searchParameterToTableName(searchParameter.type);
-
-  const rootSelect = `SELECT ${columns.join(
-    ", "
-  )} FROM ${search_table} WHERE parameter_url = $${index++} `;
-  values = [...values, searchParameter.url];
-
-  switch (searchParameter.type) {
-    case "number":
-    case "string":
-    case "uri":
-    case "date":
-    case "token":
-    case "reference":
-    case "quantity": {
-      const result = PARAMETER_FILTERS[searchParameter.type](
-        ctx,
-        parameter,
-        values
-      );
-      index = result.values.length + 1;
-      const query = `(${rootSelect} AND workspace=$${index++} AND ${result.query})`;
-
-      values = [...result.values, ctx.workspace];
-      return {
-        index,
-        values,
-        query,
-      };
-    }
-    default:
-      throw new OperationError(
-        outcomeError(
-          "not-supported",
-          `Parameter of type '${searchParameter.type}' is not yet supported.`
-        )
-      );
-  }
-}
+import { buildParameterSQL } from "./parameters/index.js";
 
 function buildParametersSQL(
   ctx: FHIRServerCTX,
