+++ packages/server/src/resourceProviders/postgres/search/parameters/reference.ts
@@ -4,7 +4,7 @@ import { FHIRServerCTX } from "../../../../ctx/types.js";
 import { searchParameterToTableName } from "../../../utilities/search/parameters.js";
 import { SearchParameterResource } from "../types.js";
 import { FilterSQLResult } from "./types.js";
-import { buildParameterSQL } from "../index.js";
+import { buildParameterSQL } from "./index.js";
 import { or } from "../../../utilities/sql.js";
 
 /*
